+++ database/migrations/2025_11_18_140350_create_player_participants_table.php
@@ -11,15 +11,19 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('player_participants', function (Blueprint $table) {
-            $table->unsignedBigInteger('event_ID');
-            $table->unsignedBigInteger('user_ID');
-            $table->integer('final_placement')->unsigned();
-            $table->primary(['event_ID','user_ID']);
-            $table->foreign('event_ID')->references('event_ID')->on('events')->onDelete('cascade');
-            $table->foreign('user_ID')->references('user_ID')->on('users')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (! Schema::hasTable("player_participants")){
+            Schema::create('player_participants', function (Blueprint $table) {
+                $table->unsignedBigInteger('event_ID');
+                $table->unsignedBigInteger('user_ID');
+                $table->integer('final_placement')->unsigned();
+                $table->timestamps();
+
+                $table->primary(['event_ID','user_ID']);
+                $table->foreign('event_ID')->references('event_ID')->on('events')->onDelete('cascade');
+                $table->foreign('user_ID')->references('user_ID')->on('users')->onDelete('cascade');
+                
+            });
+        }
     }
 
     /**
